Extract extractWebJars to buildSrc and make it configuration cache capable
