+++ graph/gradle.properties
@@ -4,6 +4,6 @@ org.gradle.configuration-cache=true
 org.gradle.configuration-cache.problems=fail
 
 org.gradle.warning.mode=fail
-org.gradle.jvmargs=-Dorg.gradle.deprecation.trace=true
+org.gradle.jvmargs=-Xmx512M -Dorg.gradle.deprecation.trace=true
 
 kotlin.stdlib.default.dependency=false
